Add unblocked_terms.txt to enable COIL pre-submit
This enables the pre-submit to check for COIL terms that do not
currently exist or are very rare in this git repo. As other terms
are cleaned from this codebase they can also be removed from this
file to enable the pre-submit for them.

See repohooks/README.md for more details.

BUG=None
TEST=None

Change-Id: I58e0f0801e8424e4763fa107a8df75f39082f6fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2660090
Commit-Queue: Bernie Thompson <bhthompson@chromium.org>
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>