gerrit: make --no-emails setting a common one
Many actions that update server settings have an option to disable
sending out notifications, so hoist that to the common parser.

Also plumb it through for a few more actions that were missing it.

This also fixes the existing --dry-run flag which was working for
subparsers, but not the main command.

BUG=None
TEST=CQ passes

Change-Id: I909f5ff68e6b588c2f663ea0046ad9484269ff0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2705104
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>