gs_unittest: simplify init patching
The CheckFile call in __init__ was made conditional a long time ago
making the need to mock it out in the tests unnecessary, so we can
delete that logic.

Instead, we should be patching the cache func so return 'gsutil' so
it doesn't try hitting the network.

BUG=chromium:1015720
TEST=`./run_tests` passes

Change-Id: I85aff83a2c0a05d570f6956699a3f8d92fb77b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2689247
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Stephane Belmon <sbelmon@google.com>
Reviewed-by: Stephane Belmon <sbelmon@google.com>