cbuildbot: fix double mocks
This unittest was patching RunTestSuite twice with autospec enabled
which doesn't really work (we need the spec of the target, not of
the mock).  It also breaks with newer versions of mock that check
the target argument more thoroughly and reject a mock.

BUG=chromium:1015720
TEST=`./run_tests` passes

Change-Id: I5bf78d5acd7c7c3c744ffb99b6b414cd01a972d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2690931
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: Navil Perez <navil@google.com>