device_imager: Initial skeleton
The first commit for device_imager.DeviceImager that is going to be used
for cros flash. In this CL we only add skeleton and minor functions like
getting the path to the image, rebooting, clearing TMP owner, etc. The
actual code for downloading and flashing image will come in follow up
CLs.

BUG=b:172212406
TEST=run_pytest lib/device_imager_unittest.py

Change-Id: I988c8d6b2b0a70e4e82ae9a5e1c67ad5cfb13190
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2714023
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>