device_imager: Add RawPartitionUpdater
This CL adds capability to update a raw partition on the device. It is
done through reading the partition from a Chromium OS image,
simultaneously compress it, and feed it to remote device through SSH. At
the same time on the device, decompress it and write it directly ot the
target partition. It can be used to update kernel or rootfs partitions.

Also added a new function to get a command for decompressing a file on a
DUT. Single threaded program should be enough for this purpose.

BUG=b:172212406
TEST=run_pytest lib/device_imager_unittest.py

Change-Id: I1fda6b6833621fff5bd5fdfa829a550d06a6948f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2714025
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>