device_imager: Add RootfsUpdater
This class is similar to KernelUpdater and in addition it does optimize
the offset and length of the root partition (based on the file system
size) and also it runs postinst on on the target root. It has capability
to revert the rootfs update too.

BUG=b:172212406
TEST=run_pytest lib/device_imager_unittest.py

Change-Id: I29ed9797a438de98fb48af59be3e3b44434cf711
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2714026
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vyshu Khota <vyshu@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>