device_imager: Enabled it using --exp-new-flash flag in cros flash
Enable device_imager by a flag --exp-new-flash in cros flash that is
defaulted to false. This allows us to expriment/dogfood with it before
flipping the switch to default true.

This new cros flash is faster than the update_engine based one. We could
reduce the cros flash time from the original 5-8 minutes to 2-4 minutes
range. A lot of inefficiencies have been removed in the new design.

BUG=b:172212406
TEST=run_pytest
TEST=cros flash <ip> xbuddy://remote/reef-release/latest
TEST=cros flash <ip> /path/to/chromiumos_test_image

Change-Id: I15883ec64a7e4ca8a2702bc6f41a102c0c30f41e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2714029
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>