device_imager: Add progress bar
We used to have a progress bar for cros flash (to remote device), but
with architecture changes that came with nebraska.py we couldn't have it
anymore. But with device_imager we can again bring out the progress
bar. We do this by looking at the stdout fd progress of the process that
writing into the rootfs. Since writing to the rootfs is the longest one,
we can based the initial stages of the progress bar on that. Then after
that we can progress forward based on the status of postinstal, reboot
and completion of flash.

BUG=b:172212406
TEST=run_pytest
TEST=cros flash <ip> xbuddy://remote/reef-release/latest
TEST=cros flash <ip> /path/to/chromiumos_test_image

Change-Id: I0053d362234f425d47e6310b9f2b1587ebb7aec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2714422
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>