cros_chrome_sdk: Try creating linker symlinks in the toolchain dir.
A NaCl build script has trouble finding a linker it likes in
recent versions of the toolchain. This should let it proceed while
we fix-up the toolchain wrappers to associate the linker binaries
with the compilers so that clang will know where to find the right
linker when asked via '-print-prog-name'.

BUG=chromium:1182224
TEST=built chrome for arm-generic in simplechrome at 13813.0.0
TEST=trying this on chrome's simple chrome bots at crrev.com/c/2720950

Change-Id: I8a9d73b794b79417eb16df835596ca104a1e7684
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2720350
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>