service: simplify determine_android_version
This func was written to accept multiple boards and assert that the
version was the same for all of them, but then only ever called with
a single board at a time.  Even the unittests patched it as if the
API only took one board.  Drop the multiple board support to make it
look like the other android funcs in here and to simplify.

BUG=chromium:1051161
TEST=CQ passes

Change-Id: I516c1b2cc1792b676836f86b047eb5e96c9b95f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2739374
Reviewed-by: Michael Mortensen <mmortensen@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>