depgraph: drop hack to rewrite unmerged blockers as merged
This code outputs packages that are being unmerged (e.g. due to
blockers) as if they're being merged.  This is a hold over from
the original parallel emerge code that wanted to generate a plan
that guaranteed ordering between the uninstall/install of the pkgs
being blocked.  It doesn't make sense anymore for our use as a
general purpose depgraph generator, and breaks when Build API
tries to look up the source ebuilds for packages that no longer
exist (but are still installed) which comes up in incremental
builds.

BUG=None
TEST=CQ passes

Change-Id: I92521a33a1950d9ba08f23cfb17e5751db90feae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2739377
Reviewed-by: Alex Klein <saklein@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>