remote_access: rework mock
Pylint is complaining about unsubscriptable-object with the CreateTunnel
return and it's not wrong: the remote_access API returns a Popen object
which is not subscriptable.  The unittests work because it mocks this
function to return a different type: a tuple.  This feels wrong and a
bit fragile, so rewrite to have it return a fake object with an API a
bit closer to Popen.

BUG=None
TEST=tests pass & are cros lint clean

Change-Id: Ia2ef5d064d51ef1647689e3c14402b31f326ede8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2746437
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>