device: Fetch board type from device if it's not specified.
If a user forgets to pass in --board in the command-line when running
cros_run_test to run tast tests in a chromium checkout, the tool
silently falls back to trying to find the tast binaries in the chroot.
Since there's no chroot in a chromium checkout, this leads to some
weird errors.

Fetching the board when the user doesn't supply it should avoid that
failure.

BUG=none
TEST=unittest
TEST=`cros_run_test --tast` locally

Change-Id: I9858d7d12d03cbb0b93812cf023c866dc24d6eb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2774600
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>