android: rename artifacts in ebuild instead
We used to rename some artifacts in the PFQ script to avoid checksum
failures due to files having the same name (b/33072485#comment7).
But in fact the renaming can be done in the ebuild directly, so do so
instead to reduce complexity in PFQ.

In addition to ebuilds, the release builder also assumes the Android
debug symbol archive is being renamed. We introduce the new path in a
backwards-compatible way, then we remove the old path once all Android
packages have been uprevved.

BUG=chromium:1074145
TEST=run_pytest
TEST=uprev android packages locally with same version, verify that
     stable ebuilds are updated but Manifest remain unchanged
     (contains checksum of renamed files)
TEST=CQ (triggers uprev due to ebuild changes)
TEST=betty-pi-arc-release-tryjob to exercise DownloadAndroidDebugSymbols

Cq-Depend: chrome-internal:3705028
Change-Id: I47694a98ead923ff4ce73ea6b11ee7b08b28ad33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2781121
Tested-by: Shao-Chuan Lee <shaochuan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Morg <morg@chromium.org>
Commit-Queue: Shao-Chuan Lee <shaochuan@chromium.org>