process_util: update signal logic for python3
The check here for handling of SIGKILL works under Python 2, but not
Python 3.  Clean it up to work under Python 3.  This only matters if
a parent is watching a child that gets killed with SIGKILL as that's
the only one that can't really be bound.

BUG=chromium:1192665
TEST=CQ passes

Change-Id: I4adbc06ca024b6c0587e45b98995d17d14c7d5eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2806227
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>