repository_unittest: reduce mock scopes
We're seeing some flakes where RmDir is being called more than we
expect.  We're not sure why that's happening, but we can probably
make this more unlikely by reducing the scope of these mocks to
the specific function we care about rather than deferring to the
tearDown phase of the tests.

BUG=chromium:1180383
TEST=tests still pass

Change-Id: Id0125791344dffde2c975080e68642c373469364
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2818178
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>