chromeos_config: Remove AUE boards from ToT config
Remove all 91 AUE devices from Tot config reference.  Replaced verifiers
with newer boards to ensure that we do not manually reference boards
that are no longer present.  In the event that a board is added back to
a milestone, Tot in this case, we will continue to build the board but
no longer reference it specifically.

BUG=chromium:1197783, chromium:1197994
TEST=`run_tests`

Change-Id: Ib1a551aea1fa08b4afbac4e6a598e29d57d4e927
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2822548
Reviewed-by: Denis Nikitin <denik@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: LaMont Jones <lamontjones@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Mike Nichols <mikenichols@chromium.org>
Tested-by: Mike Nichols <mikenichols@chromium.org>