lddtree: use ldso's --argv0 when available [reland]
Newer glibc has a --argv0 option in current releases to explicitly
set argv[0].  This is useful for the generated shell wrappers as we
need to rename it (with an .elf suffix) and that can either confuse
the output or break tools that inspect their argv[0] strictly.

This reland includes a fix for prefixing the interp with the root.

BUG=chromium:1200189
TEST=copying an ELF w/wrappers & newer glibc uses --argv0

Change-Id: I9294a73b3d3a7a275b610930938767882091660c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2834157
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>