osutils: WriteFile: improve permissions handling
If we write a file with atomic=True sudo=True, we end up resetting
the file permissions.  This is easy to miss when we're only writing
to 644 files, but easy to see when we try updating a 755 script.

Extend the logic to preserve file permission by default, and expose
the setting to allow people to force it to something else.

BUG=None
TEST=CQ passes

Change-Id: If2cddc1aae654f32e091f75a58cd14982929d6ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2842310
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>