cros_setup_toolchains: flip order for copying ELFs
We currently copy the programs, generate their wrappers, and then copy
their libs.  This means the programs are temporarily broken, but it's
not a big deal as we don't plan on running any of this before we've
finished copying everything.

This breaks when the wrapper generation needs to probe some of the libs
(namely the ldso).  Flip the order so we copy all the deps first, and
then we generate the wrapper.

BUG=chromium:1201270
TEST=`cros_setup_toolchains --create-packages` passes

Change-Id: I945b7e1879482ee0f4b085397a0022e0942729c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2845231
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>