mock: remove self-import
The mock module imports itself to access a few constants ... that it
already has access to.  This works when we stuff this path into the
PYTHONPATH search, but breaks down when we don't (which we're in the
process of cleaning up).

BUG=b:187794934
TEST=CQ passes

Change-Id: I20e832fbd6c2c68726f8af6ae55626e9d4f39a01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2896051
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chris McDonald <cjmcdonald@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>