cros_test: Gracefully Ignore missing dirs when pulling results from DUT.
One of the dirs that can contain chrome crashes (/home/chronos/crash/)
can be non-existent for tests that don't ever log the user in.

So trying to capture that dir in that case leads to errors. By
adding ignore_failures here, we can try capturing it for every test even
if it doesn't exist.

This shouldn't mask any real errors since these results are mainly log
and crash files, and are only used for manual debugging. So this
shouldn't interfere with how the results are handled on automated runs.

BUG=none
TEST=unittest

Change-Id: I63da9290a7d9063ee011e0319bfa7a5cf0bf0a79
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2903280
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>