factory: Keep symlink of bundle components
We use "zip -r" in the BuildFactoryZip and it will replace every
symbolic link with a copied of original reference file inside the
zipfile. We previously use this mechanism to save some copy operations.

Since CL:2831092 made the setup/cgpt only works if it is a symbolic link
of setup/libx64/cgpt, we change to use "zip -y -r".

BUG=b:189238341
TEST=(0) cd ~/trunk/src/scripts
     (1) ./build_images --board=$BOARD --symlink=factory_shim \
	   factory_install
     (2) ./make_netboot --board=$BOARD \
           --image_dir ~/trunk/src/build/images/$BOARD/factory_shim
     (3) cd ~/trunk/chromite
     (4) Run commands.BuildFactoryZip
     (5) unzip factory_image.zip
     (6) Check `setup/cgpt -h` outputs the help page correctly

Change-Id: I15409a68114583a8f29419670b81ec3158fb803f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2919073
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>