CreateTarball: use --hole-detection=raw for sparse files
Our disk images are large and somewhat sparse, but they're not as
sparse as they could be.  This is because we initially create a
large sparse file, write to it when building the image, and then
delete/cleanup files.  That last step frees up space, but does not
do so as sparse.  On a guado build, this could be O(hundred MB)
savings in the recovery image tarball.

BUG=None
TEST=CQ passes

Change-Id: Idfd184282d20c036492d1c769a7703f187d760a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2923885
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: George Engelbrecht <engeg@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>