cros_sdk: lib/namespaces: Exit when main pid terminates
_ReapChildren is currently waiting for all child processes to terminate.
If one of the child processes has forked off as a background or daemon
it will never terminate. These processes are also not in the foreground,
so they won't receive Ctrl+C messages.

This change makes it so that when the "main" pid exits, we exit the init
processes. Since we are running in a PID namespace when the internal
init processes exists, the OS will SIGKILL all processes in the
namespace. This prevents the cros_sdk command from hanging.

The docstring of _ReapChildren was previously wrong. It should have
actually read:
> Reap all children, then return the status of the main pid.

With this change, the code now matches the docstring.

BUG=chromium:1192665, b:189324786
TEST=cros_sdk --debug --nouse-image -- bash -c 'echo "Pre-fork"; tmux new-session -d; sleep 8; pstree -gp; bash -c "sleep 8; echo in fork; pstree -pg" &'
Verify we exit when the first bash instance terminates

Change-Id: I6c19c44efc6696cbdf4dbf7c83145a7e0bf78b72
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2933482
Tested-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>