paygen: fix python3 issue with UnofficialSignerPayloadsClient
In python3, the return value for binascii.hexlify is a byte array,
so we go to print it in our format string and end up with a value
like b'abc123' if this is passed to an external command, the shell
consumes the quotes and you end up with 'babcd123', so formatting
the filename /tmp/hash-b'abc123' ends up with /tmp/hash-babcd123,
which looks like a valid hash value but doesn't exist.

The fix for this is to decode the byte string when it's not a string.

BUG=None
TEST=unittests

Change-Id: I2cb24665e51ab49d41121847f87c84c59a24bf2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2935698
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>
Tested-by: Sean McAllister <smcallis@google.com>
Auto-Submit: Sean McAllister <smcallis@google.com>