paygen: fix suspicious decode() behavior by removing it entirely
decode() seems to work differently on different machines, possibly
due to variances in locale and encoding behavior.  Let's not use it
and instead just use the .hex() method which was added in python3.5.

This always returns a string object directly containing the hex
representation of the bytes object.

Write a unittest to ensure we don't break again due to python
formatting our hashes as binary array with a prefix on them.

BUG=None
TEST=unittests

Change-Id: I1d70b59c1a97770529e0c1e8f497c693c93370be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2940604
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Sean McAllister <smcallis@google.com>
Tested-by: Sean McAllister <smcallis@google.com>