pylint: fork bad-continuation & bad-whitespace checks
These were removed from the latest pylint version, but we still rely
on them, so grab the checks as they existed before upstream removal.
I've tried to strip down the module to only contain the code that
matters to these two checks, but it was a bit hard as it was a single
checker handling many different tests.

BUG=None
TEST=`cros lint` flags bad formatting again

Change-Id: I979cee34a3d1462edc04e7489569345e17b0ab92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3035563
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>