sdk: lookup current manifest branch in manifests.git
The current logic omits the manifest branch if m/master or m/main is
seen based on the assumption a checkout has only ever been used for
one branch.  If the dev switches between them, then we'd omit the
branch forever.

Lookup the current branch from the manifests.git repo directly to
avoid this confusion.

BUG=b:187789161
TEST=cros_sdk now shows branch with multiple m/ locals

Change-Id: Icb2f5e5aeb7a18bfb0a24755e4a961f0a646bffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2945911
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>