osutils: add Open & Chroot context managers
Rewrite the Chdir manager to use the new Open manager to make it a
bit robust.  Add a Chroot manager too as a cheap way of chrooting
into a path.

The downside is that chroot requires root access, and we don't have
anyway of getting that without sudo, and that requires a fork+exec.
So unittests for it are skipped.

BUG=b:191307774
TEST=`cros_sdk` still works

Change-Id: Ib365fe6c711fdb867e4aecf414bce29f17693ba9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2983561
Reviewed-by: Sergey Frolov <sfrolov@google.com>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>