wrapper: normalize dashes to underscores
This will allow us to have `cros-workon` as a program with the
underlying module still names "cros_workon.py".  The wrapper
today already supports "cros-workon.py", but we don't want to
rename modules, or use dashes in them in general as it makes
it a lot harder to import.

BUG=b:187792699
TEST=`cros-workon --help` passes

Change-Id: I3f2aaae1ea89534cd2db32305b216f513f35b96c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2975662
Reviewed-by: Alex Klein <saklein@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>