gs: remove GetDefaultGSUtilBin from public API
This function does two things: primarily, initialize the local gsutil
cache, and secondarily, return the path to the internal gsutil script.
We want to get away from that latter part as it complicates the code
as we have to make sure both the Python APIs work as well as skipping
us entirely and running the gsutil script.

Using GetDefaultGSUtilBin as a proxy for the first behavior is a bit
ugly, but permissible.  We've already migrated all the users of the
latter behavior.  Rename the function to reflect that intention and
drop support for returning the direct path to the script.

BUG=b:188645230
TEST=CQ passes

Change-Id: I4d7f9f7a988d096440dfb796e368189a85f64917
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2980317
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>