gs: overhaul crcmod & gsutil execution handling
Now that we've internalized all the code paths to the gsutil script,
we can change its behavior and how we execute the gsutil program.
This allows us to avoid hacking up the gsutil script just to change
its shebang.  This also allows us to make sure a compiled crcmod is
available to gsutil in more situations.

The first to do is see if the current Python runtime has a compiled
crcmod available.  If so, there's nothing for us to do, and we'll
execute gsutil with that (instead of `python3`).  This is the best
option as it's the fastest and most common in CrOS environments.

If one isn't available, we'll try compiling the copy bundled with
gsutil (again using the current Python runtime instead of python3).
The compiled result is cached, so other than the one-time cost, it
will still be fast.

If that fails, we'll try vpython if it's available.  This should
work on bots (outside of the CrOS SDK) if the active version of
Python has been injected by the bot runtime.  We run chromite in
places that don't have depot_tools, so we can't assume we always
have access to vpython.  We also take a perf hit on every run as
vpython has known startup performance issues.

If all that fails, we give up like we do today.  We'll still use
crcmod, but it won't be compiled, so it won't be as fast in some
situations.  For very large files, this could add 10-30 minutes.

BUG=b:188645230
TEST=CQ passes

Change-Id: Ie170b2165100553a5a4721609104db2db634ef80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2980320
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>