Implement base_is_recovery.
On reven the base image has a graphical installer enabled, which is used
to recover the device. Recovery images in the traditional sense can not
be booted on reven devices, which run legacy BIOS or UEFI firmware.

When base_is_recovery is set to True, we copy the base image to
recovery_image.bin instead of running mod_image_for_recovery.sh. We also
make that the images list contains a base image, because the process
would fail otherwise.

BUG=b:186631313
TEST=Run chromeos_config_unittest.py locally.

Change-Id: Ifdb4c3a2c186062ccdb849159b2173db0e6c0f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2983861
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Tested-by: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>