pytest: get pylint tests working again
The initial pytest port disabled a bunch of pylint tests since the
modules couldn't be imported.  Make pylint available to the test
environment and delete the automatic disable logic so this doesn't
regress again in the future.

Also do the same for the _emerge import.

BUG=None
TEST=`./run_tests` passes

Change-Id: Ic438b24c1e7f2c34527baca386ae537e2305f0f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/2983438
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>