gs: try creating vpython spec as non-root first
Try to create the vpython spec as the current user first.  This is
faster, and works on more systems (e.g. where sudo isn't available).
We need sudo here only if the cache was already created as root, so
we can add a bit more logic to make it dynamic.

BUG=b:188645230, chromium:1230451
TEST=CQ passes

Change-Id: I717edc04952ad993f1f5f462813a1b48b17c8eb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3039203
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>