deploy_to_termina: Resize rootfs/tools images as needed
The vm_rootfs and vm_tools images are shrunk to the minimum size to save
space (since in normal use they're read-only) but this means we have no
extra space if what we're copying over is larger. Now we add an extra
200M, copy our files, then re-shrink the image back to the new minimum.
Since truncate adds holes and we re-shrink images back to the new
minimum the overall space requirement is pretty much the same as before
and it adds <100ms to the time taken.

BUG=chromium:1229046
TEST=deploy_to_termina --board tatl eve-proxy termina_container_tools failed before, now succeeds

Change-Id: I27269d81f8dccf68434c3d6536a5983a96d1ad39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3038015
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: David Munro <davidmunro@google.com>
Tested-by: David Munro <davidmunro@google.com>