Add builders for reven-vmtest.
The reven-vmtest board is a flavour of the reven board for VM testing.
Here we disable HW tests and set paygen to false in the hopes that it
will skip signing.

We also specify a test image only, which requires us to disable the
image and signer tests: the aforementioned all run for the parent board
so we are less concerned with running them here.

BUG=chromium:1232182
TEST=./chromeos_config_unittest.py

Change-Id: I4da08939ff147c7d5b101f9414a2a5d72731c024
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3048235
Auto-Submit: Joseph Sussman <josephsussman@google.com>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Tested-by: Joseph Sussman <josephsussman@google.com>