paygen_payload_lib: less scary message when payloads wait
When payloads are behind the memory semaphore and don't yet have a
lock, which indicates permission to run, they output a message which
makes it appear there is some failure. There is not, this is normal for
paygen jobs with a lot of payloads.

BUG=b:193910799
TEST=logging only

Change-Id: Ia1ddf7cd2fd47eb54d3408c0090235599dafb4f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3052978
Tested-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>