vpython: refresh conflicting deps
Newer version of vpython seems to trigger a version check that it
wasn't previously, so these tools are failing:

ERROR: Cannot install -r /tmp/vpython_bootstrap362632330/requirements.txt (line 1)
       and wrapt==1.10.11 because these package versions have conflicting dependencies.

The conflict is caused by:
    The user requested wrapt==1.10.11
        astroid 2.4.2 depends on wrapt~=1.11

Update to the latest wrapt version to avoid this.

We also hit a conflict with google-cloud & protobuf modules in
run_tests, so we have to refresh those deps as well.  This is a
good thing as it makes the protobuf version match the vendored
copy we have in the chromite third_party/.

BUG=chromium:1230447
TEST=CQ passes

Change-Id: I96e37acf25b2939789f0d2c03ccd9ee71bfbaccf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3057055
Tested-by: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>