cros_chrome_sdk: Use the GN default value for dcheck_always_on.
Currently, DCHECKs default to false for CrOS:
https://source.chromium.org/chromium/chromium/src/+/main:build/config/dcheck_always_on.gni;drc=4da74fd850f7abbfe249b5f208c1b370535d3b3e;l=32

But ideally we remove the "!is_chromeos" bit and let it default to
true on CrOS for developers.

But to avoid bots from inadvertently picking up a default val of true
when we don't want them to, we set explicitly set dcheck_always_on to
false in the ebuild.

So this pops() dcheck_always_on from the args passed from the ebuild
to simple chrome, and lets dcheck_always_on pick up GN default val
for developers.

BUG=none
TEST=unittest

Change-Id: I81fe928c7e6b0ffbfaa815801b5ddfed1888c2db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3060533
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Tested-by: Ben Pastene <bpastene@chromium.org>