scripts: tricium clippy strips git repo from paths
This CL resolves a problem which is preventing comments from surfacing
in Gerrit for clippy findings in the build_linters recipe. Findings are
currently keeping the git repo path as a prefix, but tricium expects
this to be stripped away.

For example, we may see:
`/mnt/host/source/src/platform2/cros-fuzz/src/lib.rs`

Whereas tricium expects: `cros-fuzz/src/lib.rs`

This fix includes temporary defaulting of the git repo to
`/mnt/host/source/src/platform2` which is currently the only supported
location, but future CLs will detect this path within the build_linters
recipe and supply it to GetClippyLints in the Build API.

This CL also removes the file filter parameter which was never used and
caused problems with relative paths which are now required. This was
going to be implemented at a later date but is no longer relevant due to
other optimizations.

BUG=b:195056381, b:188578936
TEST=chromite/scripts/run_test,
./call_scripts/toolchain__get_clippy_lints, manual testing of
./tricium_cargo_clippy

Change-Id: I3e3fe685746e9e650fcf2b14d8cddd9b7d9c68c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3061330
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Jack Neus <jackneus@google.com>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>