scripts: Tricium cargo clipy locations refactor
This CL modifies the way tricium_cargo_clippy.py parses locations in
order to fix a bug that was causing the wrong filename to be attached to
some lints. The bug was due to the fact that the target field from
Clippy describes the main file, not the file from which the lints
arrise. To address this, we will now send the package location as the
first line in the input files.

We also remove redundant file path fields that were never actually being
used.

BUG=b:187795424
TEST=./run_tests with modified unit tests, and verified ./callscripts
output for the clippy lints endpoints in the toolchain service

Cq-Depend: chromium:3072619
Change-Id: I74bf907a1b764ea3bf65015078e1fd61bc92bee0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3076085
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: Jack Neus <jackneus@google.com>