portage_util.GetSourceInfo: Allow ignoring missing source paths
When the ebuild's source paths include repos that are not synced,
e.g. in the notdefault group, the method currently raises an error.
This error is not necessary, as the rest of the function can behvae
as expected when the directory does not exist.
Instead, check the directory exists in GetCommitId, which is where
the directories end up actually being used and needed.

BUG=b:196539844
TEST=./run_tests

Change-Id: I0e74f441bcf6a5282e806d7c5855e0ad62558380
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/chromite/+/3095093
Tested-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Alex Klein <saklein@chromium.org>